working on perangkat v2

+++ application/apps/perangkat/models/M_perangkat.php
@@ -7,8 +7,9 @@
    }
 
    function getdata_perangkat() {
-    $this->db->order_by('id_perangkat');
-    $data = $this->db->get('tm_perangkat');
+    $this->db->join('tm_lokasi b', 'b.id_lokasi = a.id_lokasi', 'inner');
+    $this->db->order_by('a.id_perangkat');
+    $data = $this->db->get('tm_perangkat a');
     return $data;
    }
 
@@ -36,14 +37,37 @@
     return $this->db->update('tm_perangkat', $data_update);
    }
 
-   function hapus($id) {
+   function hapus_perangkat($id) {
     $this->db->where('id_perangkat', $id);
-    return $this->db->delete('tm_perangkat');
+    if ($this->db->delete('tm_kegiatan')) {
+     $this->db->where('id_perangkat', $id);
+     return $this->db->delete('tm_perangkat');
+    }
+   }
+
+   function hapus_kegiatan($id_keg) {
+    $this->db->where('id_kegiatan', $id_keg);
+    $del = $this->db->delete('tm_kegiatan');
+    if ($del) {
+     return true;
+    } else {
+     return false;
+    }
    }
 
    function insert_kegiatan($data_kegiatan) {
     $insert = $this->db->insert('tm_kegiatan', $data_kegiatan);
     if ($insert) {
+     return true;
+    } else {
+     return false;
+    }
+   }
+
+   function update_kegiatan($id, $data_update) {
+    $this->db->where('id_kegiatan', $id);
+    $update = $this->db->update('tm_kegiatan', $data_update);
+    if ($update) {
      return true;
     } else {
      return false;
